+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeTag.scale` and `EnvelopeTemplateTag.scale`
 - Add multiple fields to `EnvelopeTemplateRecipient`
 - Add `Envelope.template`
+- Add new endpoints `MyEndpoint`, `MyDashboardEndpoint`, `MyEnvelopesEndpoint`
+- Add new resources `MyDashboard`, `MyEnvelope`, `MyEnvelopeDocument` and `MyEnvelopeRecipient`
+- Add new `MyAccount` endpoint and resource
 
 ### Modify
 - Add `final` to all resources